+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - **Breaking**: Deprecate `{Array,Group}::new` for `open`, and add `open_opt`
  - Change internal structure of various iterators to use `std::ops::Range` and remove redundant `length`
  - **Breaking**: `Indices::new_with_start_end` now takes a `range` rather than a `start` and `end`
+ - `RecommendedConcurrency::new` takes `impl std::ops::RangeBounds<usize>` instead of `std::ops::Range`
 
 ### Removed
  - **Breaking**: Remove `bytes` dependency